Move `PlanAmount` and `SetAmountViewModel` to separate files
